Rename 'ignored-countries.js' file
